5758 - Split links util
