+++ src/meta/cycleData/links/getI18nValidationStatusLabelKey.ts
@@ -1,0 +1,8 @@
+import { LinkValidationStatusCode } from 'meta/cycleData/links/link'
+
+export const getI18nValidationStatusLabelKey = (code: LinkValidationStatusCode): string => {
+  if ([LinkValidationStatusCode.success, LinkValidationStatusCode.empty].includes(code)) {
+    return `common.${code}`
+  }
+  return `admin.${code}`
+}
